Tarea 2/grpc_server/requirements.txt
grpcio
grpcio-tools
kafka-python
pandas

